pyproject.toml
[project]
name = "stock-evaluations-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "authlib>=1.6.8",
    "edgartools>=5.12.2",
    "fastapi>=0.128.0",
    "greenlet>=3.3.1",
    "httpx>=0.28.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "langchain>=1.2.7",
    "langchain-anthropic>=1.3.1",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "langchain-mistralai>=1.1.1",
    "langchain-xai>=1.2.2",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "weasyprint>=68.1",
    "yfinance>=1.1.0",
]
